+++ DaemonLinux.cpp
@@ -8,7 +8,7 @@
 #include <fcntl.h>
 #include <sys/stat.h>
 
-#include "Log.h"
+#include "util/Log.h"
 #include "util/util.h"
