+++ Datagram.cpp
@@ -2,7 +2,7 @@
 #include <vector>
 #include <cryptopp/sha.h>
 #include <cryptopp/gzip.h>
-#include "Log.h"
+#include "util/Log.h"
 #include "TunnelBase.h"
 #include "RouterContext.h"
 #include "Destination.h"
