+++ HTTPServer.cpp
@@ -2,7 +2,7 @@
 #include <boost/lexical_cast.hpp>
 #include <boost/date_time/posix_time/posix_time.hpp>
 #include "util/base64.h"
-#include "Log.h"
+#include "util/Log.h"
 #include "Tunnel.h"
 #include "TransitTunnel.h"
 #include "transport/Transports.h"
