+++ SAM.cpp
@@ -6,7 +6,7 @@
 #include <boost/lexical_cast.hpp>
 #include "util/base64.h"
 #include "Identity.h"
-#include "Log.h"
+#include "util/Log.h"
 #include "Destination.h"
 #include "ClientContext.h"
 #include "SAM.h"
