+++ TransitTunnel.cpp
@@ -1,6 +1,6 @@
 #include <string.h>
 #include "util/I2PEndian.h"
-#include "Log.h"
+#include "util/Log.h"
 #include "RouterContext.h"
 #include "I2NPProtocol.h"
 #include "Tunnel.h"
