+++ transport/SSU.cpp
@@ -1,6 +1,6 @@
 #include <string.h>
 #include <boost/bind.hpp>
-#include "Log.h"
+#include "util/Log.h"
 #include "util/Timestamp.h"
 #include "RouterContext.h"
 #include "NetDb.h"
