+++ util/Log.h
@@ -7,7 +7,7 @@
 #include <fstream>
 #include <functional>
 #include <chrono>
-#include "util/Queue.h"
+#include "Queue.h"
 
 enum LogLevel
 {
